Revert 56-bit relfilenode change and follow-up commits.
BABELFISH-CONFLICT: see Postgres community repo for original commit

There are still some alignment-related failures in the buildfarm,
which might or might not be able to be fixed quickly, but I've also
just realized that it increased the size of many WAL records by 4 bytes
because a block reference contains a RelFileLocator. The effect of that
hasn't been studied or discussed, so revert for now.

(cherry picked from commit a448e49bcbe40fb72e1ed85af910dd216d45bad8)