ci: Add 32bit build and test
It's easy enough to make changes that break on 32bit platforms and few people
test that locally. Add a test for that to CI. LLVM is disabled on 32bit
because installing it would bloat the image size further.

The debian w/ meson task is fast enough that we can afford to test both.

Use the occasion of a separate run of the tests to run them under LANG=C,
we've recently discovered there's not a lot of testing in the buildfarm for
the case.

Discussion: https://postgr.es/m/4033181.1664395633@sss.pgh.pa.us
(cherry picked from commit ccf36ea2580f66abbc37f27d8c296861ffaad9bf)