Support duplicate trigger name in different database
Previously, duplicate trigger name is not allowed in different database.
But in sql server, different database can have the same trigger name. In
this PR, we check if the trigger with the same trigger is not in current database, then the trigger can be created. And when drop trigger, we modified the pre check trigger in parsing stage to not throw error, and find the correct trigger to drop with trigger and database name. In this PR, we make ''get_object_address_trigger_tsql'' a hook type and move the function to hook.c file.

Task:BABEL-3117
Signed-off-by: Zhenye Li <zhenye@amazon.com>