Support nondeterministic collations for LIKE in CHECK constraint clause
Utility statements such as ALTER TABLE are not planned. Due to collation
like to ilike transformation logic being called in a planner hook, CHECK
constraint in utlity commands do not support LIKE for nondeterministic
collations.

To solve this problem, we add a hook to invoke the like to ilike
transformation function after the raw parsetree is transformed
into an executable expression.

Task: BABEL-3319
Signed-off-by: Avantika Dasgupta <davantik@amazon.com>