Remove unneeded include in test_slru.c
As introduced in 006b69f, the order of the headers was incorrect.
However, it happens that lwlock.h can just be dropped from the list, so
let's be clean and remove it, fixing the order of the listed headers.

(cherry picked from commit 3894d21d22eba8438743eaf453b2f2c85e479125)