Revert "Prevent instability in contrib/pageinspect's regression test."
This reverts commit 5cda142bb9d2bd7e7ed1c22ae89afe58abfa8d7b
(in v14 only).

It turns out that that fails under force_parallel_mode = regress,
because pageinspect's disk-access functions are marked parallel
safe, which they are not if you try to use them on a temp table.
The cost of fixing that pre-v15 seems to exceed the value of
making this test case fully stable, so we will just leave things
as-is in v14.

(cherry picked from commit 47a22dc2cb89aca2e54c9cf9fe2da4a5e8ba4cb4)