Fix early string free to support GUC revert (#92)
The IDENTITY_INSERT setting uses strings to store if the setting is on
or off and for which table. It is a babelfish GUC, so we store old
values for revert in the session_stack field, as opposed to stack. The
engine will check the stack to see if any string values are being used,
and if not, will free the string. This causes issues as the string is
improperly freed while used in session_stack. This change fixes the
issues, checking session_stack as well as stack before freeing.

Task: BABEL-3092

Signed-off-by: Walt Boettge <wboettge@amazon.com>