Fix scope_identity rewrite for other types
A previous commit rewrites scope_identity to return an integer so that
an index may be used for certain queries. This only applied to columns
of type integer, but identity columns can be of type smallint and bigint
as well. Now, if scope_identity is compared against columns of these
types the call will be rewritten and the index will be used.

Signed-off-by: Walt Boettge <wboettge@amazon.com>