Table Variables inside Transactions (engine code changes)
This commit refactors how Table Variables work inside a Transaction
especially when there is a rollback. The Table Variable metadata,
on-disk files, and tuples should be able to survive rollback
because Table Variables are non-transactional.

Also, Table Variables primary key, constraint and toast tables should
be non-transactional as well.

On high-level there are three main components in this commit:

  1. Metadata:
     Table Variable metadata are added into ENR in order to bypass
     catalog MVCC.

  2. On-disk files:
     Table Variable relfilenode are not added to pending_deletes
     during table creation so it would not be deleted in case of rollback.

  3. Tuples:
     Introduced a new set of MVCC, Vacuum, Toast visibility functions.

Signed-off-by: Kristian Lejao <klejao@amazon.com>
Task: BABEL-579