Fix the crash when set variable from pg endpoint using babel database
Previously when try to set variable from pg endpoint when that backend is
using babel database, the postgres backend will crash.

The root cause for this crash is pg backend when returnning will call
security definer in report Guc changes, in such condition, it's not
within a transaction.

This commit fix this issue by checking the transaction state in security
definer.

Task: BABEL-3301
Signed-off-by: Zhibai Song <szh@amazon.com>