Fix issue with UNION ALL and numeric types
Changing the query's target list to the result of the leftmost query
affected how UNION ALL works with numeric types. So that the largest
scale and precision is returned between different numeric columns when
using UNION ALL, the vartypmod field is set to -1 with UNION ALL.

However, we need to  switches the usual target list to the un-modified
leftmost query's target list, causing vartypmod to not equal -1.

This change saves and restores the old target list after processing the
ORDER BY clause. We also modify this list by filling in the ressortgroupref
fields.

Signed-off-by: Walt Boettge <wboettge@amazon.com>